discord-bot/requirements.txt
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
urllib3
aiohttp
yarl
frozenlist